✨(frontend) create hook useTransRole
Create the hook useTransRole to get a translated
role.
