+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useTransRole';
