+++ src/frontend/apps/impress/src/features/docs/doc-management/index.ts
@@ -1,4 +1,5 @@
 export * from './api';
 export * from './components';
+export * from './hooks';
 export * from './types';
 export * from './utils';
